The tests used to fail when the lxml library was installed: Error raised InvalidNameError. This commit ensures that the tests pass no matter the user has installed whichever parsing library with bs4.
**Explanation:**

When using lxml (recommended parser in BeautifulSoup4) the parsing is little different then the inbuilt parser:

    BeautifulSoup('<input name="hi">')
    # gives "<html><body><input name='hi' /></body></html>"
    # while the inbuilt parser gives "<input name='hi' />"
    # Thus the add_field function used to fail by giving InvalidNameError

The parsing of complicated forms is way much better using `lxml` and RoboBrowser compliments it well.
The only issue caused was due to the individual field parsings.
Forcing inbuilt parser for fields parsing fixes the issues with individual fields while ensuring the efficiency of lxml for parsing of forms.