Update docs for change to Rap Genius website
The Genius website (formerly Rap Genius) changed, which broke the
existing documentation example. The search results now include results
matching song title as well as artist, making it difficult to guarantee
the first song would match the artist being searched for. As a result
the example was updated to search for a more unique band name.

[Resolves #33]